renamed document data to document info
