+++ readnext/modeling/__init__.py
@@ -1,4 +1,4 @@
-from readnext.modeling.document_data import Document, DocumentScore
+from readnext.modeling.document_info import DocumentInfo, DocumentScore
 from readnext.modeling.model_data import (
     CitationModelData,
     CitationModelDataFromId,
@@ -9,7 +9,7 @@ from readnext.modeling.model_data import (
 )
 
 __all__ = [
-    "Document",
+    "DocumentInfo",
     "DocumentScore",
     "CitationModelData",
     "CitationModelDataFromId",
